+++ tests/cpp/test_slice_an_nddata.cpp
@@ -7,17 +7,18 @@
 
 #include "merlin/logger.hpp"
 #include "merlin/array/copy.hpp"
+#include "merlin/array/slice.hpp"
 
 int main(void) {
     MESSAGE("Test NdData class.\n");
-    float A_ptr[6];
-    std::iota<float *, float>(A_ptr, A_ptr + 6, 0.0);  // A = {0, ..., 5}
-    merlin::array::NdData A_array(A_ptr, 2, {2, 3}, {3*sizeof(float), sizeof(float)});
+    double A_ptr[6];
+    std::iota<double *, double>(A_ptr, A_ptr + 6, 0.0);  // A = {0, ..., 5}
+    merlin::array::NdData A_array(A_ptr, 2, {2, 3}, {3*sizeof(double), sizeof(double)});
     MESSAGE("NdData A = [[%.1f %.1f %.1f], [%.1f %.1f %.1f]].\n", A_array.data()[0], A_array.data()[1],
             A_array.data()[2], A_array.data()[3], A_array.data()[4], A_array.data()[5]);
 
-    float B_ptr[6];
-    merlin::array::NdData B_array(B_ptr, 2, {2, 3}, {3*sizeof(float), sizeof(float)});
+    double B_ptr[6];
+    merlin::array::NdData B_array(B_ptr, 2, {2, 3}, {3*sizeof(double), sizeof(double)});
     MESSAGE("Initialize NdData B.\n");
     MESSAGE("NdData B = [[%.1f %.1f %.1f], [%.1f %.1f %.1f]]\n", B_array.data()[0], B_array.data()[1],
             B_array.data()[2], B_array.data()[3], B_array.data()[4], B_array.data()[5]);
@@ -32,4 +33,5 @@ int main(void) {
     MESSAGE("Shape of C: [%" PRIu64 ", %" PRIu64 "].\n", C_array.shape()[0], C_array.shape()[1]);
     MESSAGE("NdData C = [[%.1f %.1f], [%.1f %.1f]].\n", C_array.data()[0], C_array.data()[2],
             C_array.data()[3], C_array.data()[5]);
+
 }
